+++ proto/control-api/CHANGELOG.md
@@ -6,6 +6,21 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [0.12.0] · 2024-??-?? (unreleased)
+[0.12.0]: /../../tree/medea-control-api-proto-0.12.0/proto/control-api
+
+[Diff](/../../compare/medea-control-api-proto-0.11.0...medea-control-api-proto-0.12.0)
+
+### Upgraded
+
+- Dependencies:
+    - [`medea-client-api-proto`] to `0.6` ([#151]).
+
+[#151]: /../../pull/151
+
+
+
+
 ## [0.11.0] · 2024-02-12
 [0.11.0]: /../../tree/medea-control-api-proto-0.11.0/proto/control-api